dbt_snowflake_project/requirements.txt
# Core dbt and Snowflake dependencies
dbt-core>=1.6.0,<2.0.0
dbt-snowflake>=1.6.0,<2.0.0

# Data quality and testing
great-expectations>=0.17.0,<1.0.0
# Note: dbt-expectations is installed via 'dbt deps', not pip

# Airflow dependencies (optional - install separately if needed)
# apache-airflow>=2.6.0,<3.0.0
# apache-airflow-providers-snowflake>=4.0.0,<6.0.0
# apache-airflow-providers-slack>=7.0.0,<9.0.0

# Monitoring and visualization
streamlit>=1.25.0,<2.0.0
plotly>=5.15.0,<6.0.0
pandas>=1.5.0,<3.0.0
snowflake-connector-python>=3.0.0,<4.0.0

# Development tools
sqlfluff>=2.0.0,<3.0.0
pre-commit>=3.0.0,<4.0.0
black>=23.0.0,<25.0.0
isort>=5.12.0,<6.0.0
flake8>=6.0.0,<7.0.0

# Documentation
mkdocs>=1.4.0,<2.0.0
mkdocs-material>=9.0.0,<10.0.0

# Utilities
python-dotenv>=0.19.0,<2.0.0
pyyaml>=6.0.0,<7.0.0
requests>=2.28.0,<3.0.0
